Update SyntheticBundleInstaller.java
The code could end up with a filePath that had consecutive slashes. This patch checks the bundlePath if it ends with a slash. If not it adds it and it then assumes the bundlePath always ends in a slash. I.e. this should fix it even if people would call it with a bundlepath that did not end in a slash.